+++ src/commands/prompt/types.ts
@@ -1,7 +1,3 @@
-import type { ProviderConfig } from '../../engine/providers/config.js';
-import type { Provider } from '../../engine/providers/provider.js';
-import type { ChatSession } from './ui/prompt-ui.js';
-
 export interface PromptOptions {
   /** Interactive mode */
   interactive: boolean;
@@ -23,17 +19,4 @@ export interface PromptOptions {
   creative?: boolean;
   /** Precise response style */
   precise?: boolean;
-}
-
-export type SessionFeedback = {
-  fileCostWarning?: string;
-  fileCostInfo?: string;
-  stylesWarning?: string;
-};
-
-export interface SessionContext {
-  provider: Provider;
-  config: ProviderConfig;
-  chatSession: ChatSession;
-  options: PromptOptions;
 }
